Rename workflow_version_history to version_history and fix missing parenthesis in import
