+++ dafni_cli/api/minio_api.py
@@ -7,6 +7,7 @@ from dafni_cli.consts import (
     DATA_UPLOAD_API_URL,
     DATA_DOWNLOAD_API_URL,
     DATA_DOWNLOAD_REDIRECT_API_URL,
+)
 
 from dafni_cli.api.dafni_api import (
     dafni_post_request,
